pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Dmitrii L <hexletdmitrii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^7.1.1"
pytest = "^8.3.2"
gunicorn = "^23.0.0"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
validators = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
